billParser/requirements.txt
flask==2.3.3
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
pillow==10.0.0  # For the test image generator
requests==2.31.0  # For manual testing